Removed old GG code and feature flag
