+++ conf/application.conf
@@ -136,10 +136,7 @@ microservice {
     fhdds {
       serviceRedirectUrl: "http://localhost:1118/fhdds/continue"
     }
-    government-gateway {
-      host = localhost
-      port = 8570
-    }
+
     investment-tax-relief {
       serviceRedirectUrl: "http://localhost:9637/investment-tax-relief-subscription/confirm-correspondence-address"
     }
@@ -171,6 +168,3 @@ delegated-service {
 
 logout.url = "http://localhost:9025/gg/sign-out"
 cancelRedirectUrl: "https://www.gov.uk/"
-
-# using GG is set to false; if true, EMAC is false
-features.registration.usingGG = false
