+++ test/connectors/BackLinkCacheConnectorSpec.scala
@@ -25,7 +25,6 @@ import org.scalatestplus.play.{OneServerPerSuite, PlaySpec}
 import play.api.libs.json.Json
 import play.api.test.Helpers._
 import uk.gov.hmrc.http.cache.client.{CacheMap, SessionCache}
-import utils.{BusinessCustomerFeatureSwitches, FeatureSwitch}
 
 import scala.concurrent.Future
 import uk.gov.hmrc.http.HeaderCarrier
